+++ packages/cms/lists/LegislativeYuanMember.ts
@@ -1,6 +1,11 @@
 import { list } from '@keystone-6/core'
-import { text, relationship, select } from '@keystone-6/core/fields'
-import { allowAllRoles } from './utils/access-control-list'
+import { text, relationship, select, integer } from '@keystone-6/core/fields'
+import {
+  allowAllRoles,
+  excludeReadOnlyRoles,
+  withReadOnlyRoleFieldMode,
+  hideReadOnlyRoles,
+} from './utils/access-control-list'
 import { CREATED_AT, UPDATED_AT } from './utils/common-field'
 import {
   MemberType,
@@ -119,6 +124,9 @@ const listConfigurations = list({
     note: text({
       label: '特殊說明',
     }),
+    proposalSuccessCount: integer({
+      label: '提案通過數',
+    }),
     createdAt: CREATED_AT,
     updatedAt: UPDATED_AT,
   },
@@ -130,13 +138,18 @@ const listConfigurations = list({
       initialSort: { field: 'labelForCMS', direction: 'DESC' },
       pageSize: 50,
     },
+    itemView: {
+      defaultFieldMode: withReadOnlyRoleFieldMode,
+    },
+    hideCreate: hideReadOnlyRoles,
+    hideDelete: hideReadOnlyRoles,
   },
   access: {
     operation: {
       query: allowAllRoles(),
-      create: allowAllRoles(),
-      update: allowAllRoles(),
-      delete: allowAllRoles(),
+      create: excludeReadOnlyRoles(),
+      update: excludeReadOnlyRoles(),
+      delete: excludeReadOnlyRoles(),
     },
   },
   hooks: {