+++ packages/cms/lists/utils/access-control-list.ts
@@ -2,15 +2,24 @@ import envVars from '../../environment-variables'
 import {
   BaseListTypeInfo,
   ListOperationAccessControl,
+  MaybeItemFunction,
+  MaybeSessionFunction,
 } from '@keystone-6/core/types'
+import { get } from 'lodash'
+const _ = {
+  get,
+}
 
 export const RoleEnum = {
   Owner: 'owner',
   Admin: 'admin',
   Editor: 'editor',
+  Headless: 'developer_headless_account',
 } as const
 
 type AccessOperation = 'query' | 'create' | 'update' | 'delete'
+
+const readonlyRoles: string[] = [RoleEnum.Headless]
 
 export const allowRoles =
   <Operation extends AccessOperation>(
@@ -24,11 +33,16 @@ export const allowRoles =
     if (!Array.isArray(roles)) {
       return false
     }
-    return Boolean(roles.indexOf(session?.data?.role) > -1)
+    return Boolean(roles.includes(session?.data?.role))
   }
 
 export const allowAllRoles = () => {
-  const roles = [RoleEnum.Owner, RoleEnum.Admin, RoleEnum.Editor]
+  const roles = [
+    RoleEnum.Owner,
+    RoleEnum.Admin,
+    RoleEnum.Editor,
+    RoleEnum.Headless,
+  ]
   return allowRoles(roles)
 }
 
@@ -40,5 +54,33 @@ export const denyRoles =
     if (!Array.isArray(roles)) {
       return true
     }
-    return roles.indexOf(session?.data?.role) === -1
+    return !roles.includes(session?.data?.role)
   }
+
+export const excludeReadOnlyRoles = () => {
+  return denyRoles(readonlyRoles)
+}
+
+type fieldMode = 'edit' | 'hidden' | 'read'
+export const withReadOnlyRoleFieldMode: MaybeItemFunction<
+  fieldMode,
+  BaseListTypeInfo
+> = ({ session }) => {
+  const role = _.get(session, ['data', 'role'], '')
+  if (readonlyRoles.includes(role)) {
+    return 'read'
+  } else {
+    return 'edit'
+  }
+}
+
+export const hideReadOnlyRoles: MaybeSessionFunction<
+  boolean,
+  BaseListTypeInfo
+> = ({ session }) => {
+  const role = _.get(session, ['data', 'role'], '')
+  if (readonlyRoles.includes(role)) {
+    return true
+  }
+  return false
+}