+++ packages/cms/migrations/20250402011109_add_new_columns_for_legislator_and_legislative_yuan_member/migration.sql
@@ -1,0 +1,7 @@
+-- AlterTable
+ALTER TABLE `LegislativeYuanMember` ADD COLUMN `proposalSuccessCount` INTEGER NULL;
+
+-- AlterTable
+ALTER TABLE `Legislator` ADD COLUMN `externalLink` VARCHAR(191) NOT NULL DEFAULT '',
+    ADD COLUMN `meetingTermCount` INTEGER NULL,
+    ADD COLUMN `meetingTermCountInfo` VARCHAR(191) NOT NULL DEFAULT '';